refactor(agent-client): rename stream-utils to time-out-manager and update imports
